Load custom directory of snippets from user settings
These snippets are additional to the built-in snippets loaded with
Sonic Pi.

Creating a new SnippetsLoader mimicking SampleLoader we load in the
default and any user defined snippets.
